String Abbreviate for Wordbreaking
Issues: #1721 #1722 #1723
