+++ src/NuGetGallery/Views/Packages/ManagePackageOwners.cshtml
@@ -4,7 +4,7 @@
 }
 
 
-<h1 class="page-heading">Manage Owners for Package @Model.Title</h1>
+<h1 class="page-heading">Manage Owners for Package "@Model.Title.Abbreviate(50)"</h1>
 <div class="error message" data-bind="text: message, visible: message"></div>
 
 <h2>Current Owners</h2>
